+++ src/prelude.rs
@@ -3,12 +3,12 @@
 pub use crate::{
     objects::{
         CollisionGroup, PhysicsAreaTag, PhysicsAttachment, PhysicsGarbageCollector, PhysicsHandle,
-        PhysicsJointTag, PhysicsRigidBodyTag, PhysicsShapeTag, PhysicsTag
+        PhysicsJointTag, PhysicsRigidBodyTag, PhysicsShapeTag, PhysicsTag,
     },
     servers::{
-        AreaPhysicsServerTrait, BodyMode, JointDesc, JointPhysicsServerTrait, JointPosition,
-        OverlapEvent, PhysicsWorld, RBodyPhysicsServerTrait, RigidBodyDesc, ShapeDesc,
-        ShapePhysicsServerTrait, WorldPhysicsServerTrait, ContactEvent
+        AreaPhysicsServerTrait, BodyMode, ContactEvent, JointDesc, JointPhysicsServerTrait,
+        JointPosition, OverlapEvent, PhysicsWorld, RBodyPhysicsServerTrait, RigidBodyDesc,
+        ShapeDesc, ShapePhysicsServerTrait, WorldPhysicsServerTrait,
     },
     PhysicsTime,
 };